+++ src/index.js
@@ -1,16 +1,13 @@
 import { Bot } from './classes/Bot.js';
 import { ImportCommands } from './functions/importCommands.js';
 import { TextCommand } from './classes/textCommand.js';
-import { Event } from './classes/event.js';
 import { SlashCommand } from './classes/slashCommand.js';
 import checkUpdate from './functions/checkUpdate.js';
-import { EmbedBuilder } from 'discord.js';
-import { ActionRowBuilder } from 'discord.js';
-import { ButtonBuilder } from 'discord.js';
-import { StringSelectMenuBuilder  } from 'discord.js';
+import { EmbedBuilder, ActionRowBuilder, ButtonBuilder, StringSelectMenuBuilder } from 'discord.js';
 import { ImportSlashCommands } from './functions/importSlashCommands.js';
 import { ImportEvents } from './functions/importEvents.js';
 import { PermissionsBitField } from 'discord.js';
+import { Event } from './classes/event.js';
 import { Modal } from './classes/Modal.js';
 
 checkUpdate();
@@ -55,4 +52,5 @@ const parrot = {
     ImportEvents,
     Event,
 };
+
 export default parrot;